// apps/api-journeys-postgres/src/__generated__/types.ts
export type Maybe<T> = T | null
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> }
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> }
export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>
export type RequireFields<T, K extends keyof T> = {
  [X in Exclude<keyof T, K>]?: T[X]
} &
  { [P in K]-?: NonNullable<T[P]> }
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
}

export type Action = {
  gtmEventName?: Maybe<Scalars['String']>
}

export type Block = {
  id: Scalars['ID']
  parentBlockId?: Maybe<Scalars['ID']>
}

export type ButtonBlock = Block & {
  __typename?: 'ButtonBlock'
  action?: Maybe<Action>
  color?: Maybe<ButtonColor>
  endIcon?: Maybe<Icon>
  id: Scalars['ID']
  label: Scalars['String']
  parentBlockId?: Maybe<Scalars['ID']>
  size?: Maybe<ButtonSize>
  startIcon?: Maybe<Icon>
  variant?: Maybe<ButtonVariant>
}

export type ButtonColor = 'error' | 'inherit' | 'primary' | 'secondary'

export type ButtonSize = 'large' | 'medium' | 'small'

export type ButtonVariant = 'contained' | 'text'

export type CardBlock = Block & {
  __typename?: 'CardBlock'
  /** backgroundColor should be a HEX color value e.g #FFFFFF for white. */
  backgroundColor?: Maybe<Scalars['String']>
  /**
   * coverBlockId is present if a child block should be used as a cover.
   * This child block should not be rendered normally, instead it should be used
   * as a background. Blocks are often of type ImageBlock or VideoBlock.
   */
  coverBlockId?: Maybe<Scalars['ID']>
  /**
   * fullscreen should control how the coverBlock is displayed. When fullscreen
   * is set to true the coverBlock Image should be displayed as a blur in the
   * background.
   */
  fullscreen: Scalars['Boolean']
  id: Scalars['ID']
  parentBlockId?: Maybe<Scalars['ID']>
  /**
   * themeMode can override journey themeMode. If nothing is set then use
   * themeMode from journey
   */
  themeMode?: Maybe<ThemeMode>
  /**
   * themeName can override journey themeName. If nothing is set then use
   * themeName from journey
   */
  themeName?: Maybe<ThemeName>
}

export type GridAlignItems = 'baseline' | 'center' | 'flexEnd' | 'flexStart'

export type GridContainerBlock = Block & {
  __typename?: 'GridContainerBlock'
  alignItems: GridAlignItems
  direction: GridDirection
  id: Scalars['ID']
  justifyContent: GridJustifyContent
  parentBlockId?: Maybe<Scalars['ID']>
  spacing: Scalars['Int']
}

export type GridDirection = 'column' | 'columnReverse' | 'row' | 'rowReverse'

export type GridItemBlock = Block & {
  __typename?: 'GridItemBlock'
  id: Scalars['ID']
  lg: Scalars['Int']
  parentBlockId?: Maybe<Scalars['ID']>
  sm: Scalars['Int']
  xl: Scalars['Int']
}

export type GridJustifyContent = 'center' | 'flexEnd' | 'flexStart'

export type Icon = {
  __typename?: 'Icon'
  color?: Maybe<IconColor>
  name: IconName
  size?: Maybe<IconSize>
}
// ...
export type ImageBlock = Block & {
  __typename?: 'ImageBlock'
  alt: Scalars['String']
  /**
   * blurhash is a compact representation of a placeholder for an image.
   * Find a frontend implementation at https://github.com/woltapp/blurhash
   */
  blurhash: Scalars['String']
  height: Scalars['Int']
  id: Scalars['ID']
  parentBlockId?: Maybe<Scalars['ID']>
  src: Scalars['String']
  width: Scalars['Int']
}

export type ImageBlockCreateInput = {
  alt: Scalars['String']
  /** ID should be unique Response UUID (Provided for optimistic mutation result matching) */
  id?: Maybe<Scalars['ID']>
  journeyId: Scalars['ID']
  parentBlockId?: Maybe<Scalars['ID']>
  src: Scalars['String']
}

export type Journey = {
  __typename?: 'Journey'
  blocks?: Maybe<Array<Block>>
  description?: Maybe<Scalars['String']>
  id: Scalars['ID']
  locale: Scalars['String']
  primaryImageBlock?: Maybe<ImageBlock>
  published: Scalars['Boolean']
  slug: Scalars['String']
  themeMode: ThemeMode
  themeName: ThemeName
  title: Scalars['String']
}

export type JourneyCreateInput = {
  description?: Maybe<Scalars['String']>
  /**
   * ID should be unique Response UUID
   * (Provided for optimistic mutation result matching)
   */
  id?: Maybe<Scalars['ID']>
  locale?: Maybe<Scalars['String']>
  /**
   * Slug should be unique amongst all journeys
   * (server will throw BAD_USER_INPUT error if not)
   */
  slug: Scalars['String']
  themeMode?: Maybe<ThemeMode>
  themeName?: Maybe<ThemeName>
  title: Scalars['String']
}

export type JourneyUpdateInput = {
  description?: Maybe<Scalars['String']>
  id: Scalars['ID']
  locale?: Maybe<Scalars['String']>
  primaryImageBlockId?: Maybe<Scalars['ID']>
  slug?: Maybe<Scalars['String']>
  themeMode?: Maybe<ThemeMode>
  themeName?: Maybe<ThemeName>
  title?: Maybe<Scalars['String']>
}

export type LinkAction = Action & {
  __typename?: 'LinkAction'
  gtmEventName?: Maybe<Scalars['String']>
  target?: Maybe<Scalars['String']>
  url: Scalars['String']
}

export type Mutation = {
  __typename?: 'Mutation'
  imageBlockCreate: ImageBlock
  journeyCreate: Journey
  journeyPublish?: Maybe<Journey>
  journeyUpdate: Journey
  radioQuestionResponseCreate: RadioQuestionResponse
  signUpResponseCreate: SignUpResponse
  videoResponseCreate: VideoResponse
}
// ...
export type NavigateAction = Action & {
  __typename?: 'NavigateAction'
  gtmEventName?: Maybe<Scalars['String']>
}

export type NavigateToBlockAction = Action & {
  __typename?: 'NavigateToBlockAction'
  blockId: Scalars['String']
  gtmEventName?: Maybe<Scalars['String']>
}

export type NavigateToJourneyAction = Action & {
  __typename?: 'NavigateToJourneyAction'
  gtmEventName?: Maybe<Scalars['String']>
  journey?: Maybe<Journey>
  journeyId: Scalars['String']
}

export type Query = {
  __typename?: 'Query'
  journey?: Maybe<Journey>
  journeys: Array<Journey>
}

export type QueryJourneyArgs = {
  id: Scalars['ID']
  idType?: Maybe<IdType>
}

export type RadioOptionBlock = Block & {
  __typename?: 'RadioOptionBlock'
  action?: Maybe<Action>
  id: Scalars['ID']
  label: Scalars['String']
  parentBlockId?: Maybe<Scalars['ID']>
}

export type RadioQuestionBlock = Block & {
  __typename?: 'RadioQuestionBlock'
  description?: Maybe<Scalars['String']>
  id: Scalars['ID']
  label: Scalars['String']
  parentBlockId?: Maybe<Scalars['ID']>
}

export type RadioQuestionResponse = Response & {
  __typename?: 'RadioQuestionResponse'
  block?: Maybe<RadioQuestionBlock>
  id: Scalars['ID']
  radioOptionBlockId: Scalars['ID']
  userId: Scalars['ID']
}

export type RadioQuestionResponseCreateInput = {
  blockId: Scalars['ID']
  /** ID should be unique Response UUID (Provided for optimistic mutation result matching) */
  id?: Maybe<Scalars['ID']>
  radioOptionBlockId: Scalars['ID']
}

export type Response = {
  id: Scalars['ID']
  userId: Scalars['ID']
}

export type SignUpBlock = Block & {
  __typename?: 'SignUpBlock'
  action?: Maybe<Action>
  id: Scalars['ID']
  parentBlockId?: Maybe<Scalars['ID']>
  submitIcon?: Maybe<Icon>
  submitLabel?: Maybe<Scalars['String']>
}

export type SignUpResponse = Response & {
  __typename?: 'SignUpResponse'
  block?: Maybe<SignUpBlock>
  email: Scalars['String']
  id: Scalars['ID']
  name: Scalars['String']
  userId: Scalars['ID']
}
// ...
export type VideoArclight = VideoContent & {
  __typename?: 'VideoArclight'
  languageId: Scalars['String']
  mediaComponentId: Scalars['String']
  src: Scalars['String']
}
// ...
export type VideoContent = {
  src: Scalars['String']
}

export type VideoGeneric = VideoContent & {
  __typename?: 'VideoGeneric'
  src: Scalars['String']
}

export type VideoResponse = Response & {
  __typename?: 'VideoResponse'
  block?: Maybe<VideoBlock>
  id: Scalars['ID']
  position?: Maybe<Scalars['Int']>
  state: VideoResponseStateEnum
  userId: Scalars['ID']
}
// ...
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | ResolverWithResolve<TResult, TParent, TContext, TArgs>
// ...
export type TypeResolveFn<TTypes, TParent = {}, TContext = {}> = (
  parent: TParent,
  context: TContext,
  info: GraphQLResolveInfo
) => Maybe<TTypes> | Promise<Maybe<TTypes>>
// ...
export type ResolversTypes = {
  Action:
    | ResolversTypes['LinkAction']
    | ResolversTypes['NavigateAction']
    | ResolversTypes['NavigateToBlockAction']
    | ResolversTypes['NavigateToJourneyAction']
  Block: ResolverTypeWrapper<BlockType>
  Boolean: ResolverTypeWrapper<Scalars['Boolean']>
  ButtonBlock: ResolverTypeWrapper<BlockType>
  ButtonColor: ButtonColor
  ButtonSize: ButtonSize
  ButtonVariant: ButtonVariant
  CardBlock: ResolverTypeWrapper<BlockType>
  Float: ResolverTypeWrapper<Scalars['Float']>
  GridAlignItems: GridAlignItems
  GridContainerBlock: ResolverTypeWrapper<BlockType>
  GridDirection: GridDirection
  GridItemBlock: ResolverTypeWrapper<BlockType>
  GridJustifyContent: GridJustifyContent
  ID: ResolverTypeWrapper<Scalars['ID']>
  Icon: ResolverTypeWrapper<Icon>
  IconColor: IconColor
  IconName: IconName
  IconSize: IconSize
  IdType: IdType
  ImageBlock: ResolverTypeWrapper<BlockType>
  ImageBlockCreateInput: ImageBlockCreateInput
  Int: ResolverTypeWrapper<Scalars['Int']>
  Journey: ResolverTypeWrapper<JourneyType>
  JourneyCreateInput: JourneyCreateInput
  JourneyUpdateInput: JourneyUpdateInput
  LinkAction: ResolverTypeWrapper<LinkAction>
  Mutation: ResolverTypeWrapper<{}>
  NavigateAction: ResolverTypeWrapper<NavigateAction>
  NavigateToBlockAction: ResolverTypeWrapper<NavigateToBlockAction>
  NavigateToJourneyAction: ResolverTypeWrapper<
    Omit<NavigateToJourneyAction, 'journey'> & {
      journey?: Maybe<ResolversTypes['Journey']>
    }
  >
  Query: ResolverTypeWrapper<{}>
  RadioOptionBlock: ResolverTypeWrapper<BlockType>
  RadioQuestionBlock: ResolverTypeWrapper<BlockType>
  RadioQuestionResponse: ResolverTypeWrapper<ResponseType>
  RadioQuestionResponseCreateInput: RadioQuestionResponseCreateInput
  Response:
    | ResolversTypes['RadioQuestionResponse']
    | ResolversTypes['SignUpResponse']
    | ResolversTypes['VideoResponse']
  SignUpBlock: ResolverTypeWrapper<BlockType>
  SignUpResponse: ResolverTypeWrapper<ResponseType>
  SignUpResponseCreateInput: SignUpResponseCreateInput
  StepBlock: ResolverTypeWrapper<BlockType>
  String: ResolverTypeWrapper<Scalars['String']>
  ThemeMode: ThemeMode
  ThemeName: ThemeName
  TypographyAlign: TypographyAlign
  TypographyBlock: ResolverTypeWrapper<BlockType>
  TypographyColor: TypographyColor
  TypographyVariant: TypographyVariant
  VideoArclight: ResolverTypeWrapper<VideoArclight>
  VideoBlock: ResolverTypeWrapper<BlockType>
  VideoContent: ResolversTypes['VideoArclight'] | ResolversTypes['VideoGeneric']
  VideoGeneric: ResolverTypeWrapper<VideoGeneric>
  VideoResponse: ResolverTypeWrapper<ResponseType>
  VideoResponseCreateInput: VideoResponseCreateInput
  VideoResponseStateEnum: VideoResponseStateEnum
  VideoTriggerBlock: ResolverTypeWrapper<BlockType>
}

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Action:
    | ResolversParentTypes['LinkAction']
    | ResolversParentTypes['NavigateAction']
    | ResolversParentTypes['NavigateToBlockAction']
    | ResolversParentTypes['NavigateToJourneyAction']
  Block: BlockType
  Boolean: Scalars['Boolean']
  ButtonBlock: BlockType
  CardBlock: BlockType
  Float: Scalars['Float']
  GridContainerBlock: BlockType
  GridItemBlock: BlockType
  ID: Scalars['ID']
  Icon: Icon
  ImageBlock: BlockType
  ImageBlockCreateInput: ImageBlockCreateInput
  Int: Scalars['Int']
  Journey: JourneyType
  JourneyCreateInput: JourneyCreateInput
  JourneyUpdateInput: JourneyUpdateInput
  LinkAction: LinkAction
  Mutation: {}
  NavigateAction: NavigateAction
  NavigateToBlockAction: NavigateToBlockAction
  NavigateToJourneyAction: Omit<NavigateToJourneyAction, 'journey'> & {
    journey?: Maybe<ResolversParentTypes['Journey']>
  }
  Query: {}
  RadioOptionBlock: BlockType
  RadioQuestionBlock: BlockType
  RadioQuestionResponse: ResponseType
  RadioQuestionResponseCreateInput: RadioQuestionResponseCreateInput
  Response:
    | ResolversParentTypes['RadioQuestionResponse']
    | ResolversParentTypes['SignUpResponse']
    | ResolversParentTypes['VideoResponse']
  SignUpBlock: BlockType
  SignUpResponse: ResponseType
  SignUpResponseCreateInput: SignUpResponseCreateInput
  StepBlock: BlockType
  String: Scalars['String']
  TypographyBlock: BlockType
  VideoArclight: VideoArclight
  VideoBlock: BlockType
  VideoContent:
    | ResolversParentTypes['VideoArclight']
    | ResolversParentTypes['VideoGeneric']
  VideoGeneric: VideoGeneric
  VideoResponse: ResponseType
  VideoResponseCreateInput: VideoResponseCreateInput
  VideoTriggerBlock: BlockType
}
// ...
export type VideoContentResolvers<
  ContextType = GraphQLModules.Context,
  ParentType extends ResolversParentTypes['VideoContent'] = ResolversParentTypes['VideoContent']
> = {
  __resolveType: TypeResolveFn<
    'VideoArclight' | 'VideoGeneric',
    ParentType,
    ContextType
  >
  src?: Resolver<ResolversTypes['String'], ParentType, ContextType>
}